# Gamificator/Teaching-HEIGVD-AMT-2016-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Gamification platform
  description: |
    **API specification of the Gamification platform**

    You can try all HTTP operation described in this Swagger spec.
    
    Find source code of this API [here](https://github.com/sarrab/GamifyApp.git)
host: localhost:8090
basePath: /api
schemes:
  - http
  - https
paths:
...
  /users/{id}:
    get:
      description:  Retrieve one user by id
      operationId: findUserById
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - name: id
          in: path
          description: id of the user to fetch
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description:
              Returns the specified user state
          schema:
            $ref: '#/definitions/User'  
...
  /endUser:
    get:
      description: Retourne tous les utilisateurs de l'application
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
      responses:
        200:
          description:
              liste des utilisateurs
          schema:
            type: array
            items:
             $ref: '#/definitions/EndUserDTO'
    post:
      description: crée un utilisateur
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: body
          name: body
          description: données de l'utilisateur
          schema:
            $ref: '#/definitions/EndUserDTO'
          required: true
      responses:
        201:
          description:
            OK utilisateur créé
          schema:
            $ref: '#/definitions/EndUserDTO'
  /endUser/{endUser'ID}/reputation:
    get:
      description: Retourne tous les utilisateurs de l'application
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: endUser'ID
          description: user's id
          required: true
          type: integer
          format: int64 
          
      responses:
        200:
          description:
              liste des utilisateurs
          schema:
            $ref: '#/definitions/EndUserReputationDTO'
  /endUser/{endUser'ID}/badges:
    get:
      description: Retourne tous les utilisateurs de l'application
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: endUser'ID
          description: endUser'ID
          required: true
          type: integer
          format: int64 
        
      responses:
        200:
          description:
              liste des utilisateurs
  /endUser/{id}:
    get:
      description: Retourne l'utilisateur dont l'id est spécifié dans le path
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: id
          description: user's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Ok, user's details
          schema:
            $ref: '#/definitions/EndUserDTO'
    put:
      description: Apporte des modifications sur un utilisateur
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: id
          description: user's id
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: Update of a user
          schema:
            $ref: '#/definitions/EndUserDTO'
          required: false
      responses:
        200:
          description: OK, user updated
          
    delete:
      description: Supprime un utilisateur
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: id
          description: user's id
          required: true
          type: integer
          format: int64     
      responses:
        200:
          description: OK, user deleted
definitions:
  EndUserDTO:
    type: object
    properties:
      name:
        type: string
        description: Nom de l'utilisateur
...
  BadgeDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        description: Badge name.
      description:
        type: string
        description: Description du badge
      imageURI:
        type: string
        format: uri
        description: Image du badge
      application:
        type: integer
        format: int64
        description: Identifient de l'application à laquelle appartient le badge
...
  User:
    type: object
    properties:
      userId:
        type: string
      numberOfEvents:
        type: integer
...
  EndUserReputationDTO:
    type: object
    properties:
      badge:
          type: array
          items:
            $ref: '#/definitions/BadgeDTO'
      points:
        type: integer
        format: int64
        description: Nombre de points attribués
parameters:
  TokenHeader:
    description: token that identifies the app sending the request
    name: X-Gamification-Token
    in: header
    required: true
    type: string